requirements.txt
admin==0.0.1
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
Django==3.2.15
django-cors-headers==3.13.0
djangorestframework==3.14.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
msgpack==1.0.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2022.2.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.3.0
zope.interface==5.4.0
gunicorn==20.1.0
django-background-tasks==1.2.5
django-compat-1.0.15
